Add HTTP calls to divert and deliver
Minitest uses a funny way to mock. We create a MiniTest::Mock, expect the :call method to be called with ['some args'] and return <some value>. Then we stub a class method and have it return the mock. If the second argument to stub responds to #call, then #call is sent to the method with ['some args']. That's why this works.