Add source class
Acts similar to a transform. Only difference being an initial load of some meta data, like credentials and the delivery address and anything else needed later on down the pipeline.

In the future, maybe that stuff can be handled by some sort of PipelineManager and the Source can be even more similar to a transform.